# Optimized Disable Tracker to use custom list type