--- a/modules/mono/glue/GodotSharp/Benchmarks/GodotSharpBenchmarks.csproj
+++ b/modules/mono/glue/GodotSharp/Benchmarks/GodotSharpBenchmarks.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GodotSharp\GodotSharp.csproj" />
+  </ItemGroup>
+
+</Project>