--- a/modules/mono/glue/GodotSharp/Benchmarks/Program.cs
+++ b/modules/mono/glue/GodotSharp/Benchmarks/Program.cs
@@ -1,0 +1,3 @@
+using BenchmarkDotNet.Running;
+
+BenchmarkRunner.Run<BenchmarkObjectCollection>();