--- a/modules/mono/glue/GodotSharp/GodotSharp/Core/GodotObject.base.cs
+++ b/modules/mono/glue/GodotSharp/GodotSharp/Core/GodotObject.base.cs
@@ -16,7 +16,7 @@ namespace Godot
         internal IntPtr NativePtr;
         private bool _memoryOwn;
 
-        private WeakReference<GodotObject>? _weakReferenceToSelf;
+        private DisposablesTracker.Element<WeakReference<GodotObject>>? _weakReferenceToSelf;
 
         /// <summary>
         /// Constructs a new <see cref="GodotObject"/>.