This is to reduce the load of removing and loading objects 
```

BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4046/22H2/2022Update)
AMD Ryzen 7 5800X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.200
  [Host]   : .NET 6.0.27 (6.0.2724.6912), X64 RyuJIT AVX2 [AttachedDebugger]
  .NET 6.0 : .NET 6.0.27 (6.0.2724.6912), X64 RyuJIT AVX2
  .NET 8.0 : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2


```
| Method                       | Job      | Runtime  | NumberOfObjects | Mean         | Error      | StdDev     |
|----------------------------- |--------- |--------- |---------------- |-------------:|-----------:|-----------:|
| **Add_Object_Collection_Based**  | **.NET 6.0** | **.NET 6.0** | **100**             |     **9.242 μs** |  **0.0842 μs** |  **0.0746 μs** |
| Add_Object_Custom_Type_Based | .NET 6.0 | .NET 6.0 | 100             |     1.397 μs |  0.0162 μs |  0.0152 μs |
| Add_Object_Collection_Based  | .NET 8.0 | .NET 8.0 | 100             |     7.141 μs |  0.0840 μs |  0.0786 μs |
| Add_Object_Custom_Type_Based | .NET 8.0 | .NET 8.0 | 100             |     1.229 μs |  0.0202 μs |  0.0189 μs |
| **Add_Object_Collection_Based**  | **.NET 6.0** | **.NET 6.0** | **1000**            |    **91.047 μs** |  **0.5011 μs** |  **0.4442 μs** |
| Add_Object_Custom_Type_Based | .NET 6.0 | .NET 6.0 | 1000            |    13.904 μs |  0.2462 μs |  0.2303 μs |
| Add_Object_Collection_Based  | .NET 8.0 | .NET 8.0 | 1000            |    76.954 μs |  0.4841 μs |  0.4291 μs |
| Add_Object_Custom_Type_Based | .NET 8.0 | .NET 8.0 | 1000            |    12.525 μs |  0.1769 μs |  0.1655 μs |
| **Add_Object_Collection_Based**  | **.NET 6.0** | **.NET 6.0** | **10000**           | **1,200.993 μs** | **15.6022 μs** | **13.8309 μs** |
| Add_Object_Custom_Type_Based | .NET 6.0 | .NET 6.0 | 10000           |   146.205 μs |  1.2649 μs |  1.1213 μs |
| Add_Object_Collection_Based  | .NET 8.0 | .NET 8.0 | 10000           | 1,093.065 μs | 21.7546 μs | 27.5125 μs |
| Add_Object_Custom_Type_Based | .NET 8.0 | .NET 8.0 | 10000           |   135.275 μs |  1.5845 μs |  1.4046 μs |